Added the validations

+++ Controllers/CommentController.cs
@@ -23,6 +23,9 @@ namespace net_tutorial.Controllers
 
         [HttpGet]
         public async Task<IActionResult> GetAll(){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
             var comments = await _commentRepo.GetAllAsync();
             var CommentDto = comments.Select(s => s.ToCommentDto());
             return Ok(CommentDto);
@@ -37,6 +40,11 @@ namespace net_tutorial.Controllers
         }
         [HttpPost("{stockId:int}")]
         public async Task<IActionResult> Create([FromRoute] int stockId, CreateCommentDto commentDto){
+            
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            
             if(!await _stockRepo.StockExists(stockId)){
                 return BadRequest("Stock does not exists");
             }
@@ -46,6 +54,10 @@ namespace net_tutorial.Controllers
         }
         [HttpPut("{id:int}")]
         public async Task<IActionResult> Update(int id,[FromBody] UpdateCommentRequestDto updateDto){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            
             var comment = await _commentRepo.UpdateAsync(id, updateDto.ToCommentFromUpdate(id));
             if(comment == null){
                 return NotFound("Comment Not Found");
@@ -54,6 +66,10 @@ namespace net_tutorial.Controllers
         }
         [HttpDelete("{id:int}")]
         public async Task<IActionResult> Delete(int id){
+            if(!ModelState.IsValid){
+                return BadRequest(ModelState);
+            }
+            
             var commentModel = await _commentRepo.DeleteAsync(id);
             if(commentModel == null){
                 return NotFound("Comment Does not exixts");
